@@ -2,7 +2,7 @@ name: google_sign_in_web
 description: Flutter plugin for Google Sign-In, a secure authentication system
   for signing in with a Google account on Android, iOS and Web.
 homepage: https://github.com/flutter/plugins/tree/master/packages/google_sign_in/google_sign_in_web
-version: 0.9.2+1
+version: 0.10.0
 
 flutter:
   plugin:
@@ -12,23 +12,19 @@ flutter:
         fileName: google_sign_in_web.dart
 
 dependencies:
-  google_sign_in_platform_interface: ^1.1.0
+  google_sign_in_platform_interface: ^2.0.0-0
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  meta: ^1.1.7
-  js: ^0.6.1
+  meta: ^1.3.0
+  js: ^0.6.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  google_sign_in: ^4.0.14
-  pedantic: ^1.8.0
-  mockito: ^4.1.1
-  integration_test:
-    path: ../../integration_test
+  pedantic: ^1.10.0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.4"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
