@@ -15,7 +15,10 @@ import 'src/test_utils.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  gapiUrl = toBase64Url(gapi_mocks.auth2InitSuccess(GoogleSignInUserData()));
+  gapiUrl = toBase64Url(gapi_mocks.auth2InitSuccess(GoogleSignInUserData(
+    email: 'test@test.com',
+    id: '1234'
+  )));
 
   testWidgets('Plugin is initialized after GAPI fully loads and init is called',
       (WidgetTester tester) async {
